backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.1
asyncpg==0.30.0
aiosqlite==0.20.0

# Vector Database
faiss-cpu==1.9.0
qdrant-client==1.12.1

# AI / ML
langchain==0.3.14
langgraph==0.2.60
langchain-community==0.3.14
langchain-openai==0.3.0
transformers==4.47.1
torch==2.5.1
sentence-transformers==3.3.1
numpy==1.26.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1

# Monitoring / Telemetry
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-prometheus==0.50b0
prometheus-client==0.21.1

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1
jinja2==3.1.5
aiofiles==24.1.0
orjson==3.10.13
python-dateutil==2.9.0
apscheduler==3.10.4

# Notifications
aiosmtplib==3.0.2
slack-sdk==3.34.0

# PDF Reports
weasyprint==63.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
